+++ application/views/admin/themes/template_edit.php
@@ -6,9 +6,10 @@
 		<div class="bheadl"></div>
 		<div class="bheadr"></div>
 
-		<h2><?php echo _('Edit blocks') . ': ' . $template; ?></h2>
+		<h2><?php echo _('Edit blocks') . ': ' . rtrim($template, '.php'); ?></h2>
 
 		<ul>
+			<li><img class="middle" src="<?php echo site_url(THEMESPATH.'admin/widgets/icns/arrow_left.png'); ?>" /> <a href="<?php echo admin_url('themes/theme/' . $theme)?>"><?php echo _('Back to templates list'); ?></a></li>
 		</ul>
 	</div>
 
@@ -19,9 +20,12 @@
 		echo $this->view->get_messages();
 		?>
 
-		<?php foreach ($blocks as $block_name => $sections) { ?>
+		<?php
+		if (is_array($blocks) && count($blocks))
+		{
+		foreach ($blocks as $block_name => $sections) { ?>
 
-		<div class="theme_block" data-name="<?php echo $block_name; ?>"><h3><?php echo $block_name; ?></h3>
+		<div class="theme_block" data-name="<?php echo $block_name; ?>"><h3><?php echo ucfirst($block_name); ?></h3>
 
 		<?php if (is_array($sections) && count($sections)) {
 				foreach ($sections as $section) {
@@ -39,13 +43,17 @@
 		<?php } ?>
 
 
-		<br /><br /><input type="submit" class="submit long" value="<?php echo _('Save changes'); ?>" />
+			<input type="submit" class="submit long" value="<?php echo _('Save changes'); ?>" />
 		<?php
+		} else {
+			echo '<div class="message warning">' . _('This templates has no blocks.') . '</div>';
+		}
+
 
 		echo form_close();
 		?>
-
-		<br /><br />
+		<br />
+		
 
 	</div>
 
@@ -54,7 +62,10 @@
 </div>
 
 <div class="hidden">
-		<div id="add_section" style="padding:10px; background:#fff;">
+		<div id="add_section">
+		
+			<?php echo $this->view->render('admin/themes/section_composer'); ?>
+
 
 		</div>
 </div>
@@ -63,6 +74,6 @@
 <script type="text/javascript" src="<?php echo site_url(THEMESPATH.'admin/js/jquery.colorbox.js'); ?>"></script>
 <script type="text/javascript">
 $(document).ready(function() {
-	$(".link").colorbox({width:"50%", inline:true, href:"#add_section"});
+	$(".add_section").colorbox({width:"65%", inline:true, href:"#add_section"});
 });
 </script>