+++ themes/admin/js/custom.js
@@ -327,10 +327,8 @@ var bancha = {
 	},
 	blocks : {
 		_last_section : false,
-		add_section : function(el) {
-			bancha.blocks._last_section = $(this).parent('.theme_block');
-			var el_name = bancha.blocks._last_section.attr('data-name');
-			//TODO: open the popup
+		save_section : function(el) {
+			
 		}
 	}
 }